+++ sw/AMI/driver/amc_proxy.h
@@ -155,7 +155,7 @@ struct amc_proxy_pdi_download_request {
 	uint32_t partition;
 	uint16_t last_chunk;
 	uint16_t chunk;
-	uint16_t chunk_size;
+	uint32_t chunk_size;
 };
 
 /**
